lab 3: Define max digits in scoreboard and prevent overflows.
Created a function that calculates the max number available for the bullet and enemies hit counters from the given amount of digits. It is then used to reset the counter when it reaches max.

Additionally char definitions have been moved into the draw_score function and definitions of SCOREBOARD_WIDTH and SCOREBOARD_HEIGHT had errors in their names corrected.