lab 3: Add multiplier to the Targets structure.
This allows us to simplify the code by using an already existing loop to generate and save the multipliers.
In addition to this the srand() function has been moved closer to the above mentioned loop.
The comment under the defines is a personal note of an idea to be implemented.