lab 3: Add the possibility to have multiple explosions happening at the same time.
This is a fix to a problem noticed in commit 'c69c58f185988bfc7b833ae2b7c80973ec2559c4'. To do this 'is_exploding', 'x_boom' and 'y_boom' variables have been added to the Target struct.
Approperiate for loops set .is_expolding to true when a target is his and disable it when the explosion_frame_counter goes to 0. In addition to this when a target is hit 'x_boom' and 'y_boom' save the explosion location.