lab 4: Implement moving cannon.
The cannon now moves along the x axis and the barrel is stationary.

The angle variable has been renamed into cannon_position and fire_angle into fire_position.

The delta_angle const has been removed as there is no need for it now.

The x, y parameters of the barrel now set it to permanently point straight up.

The draw_scene now accepts cannon_position to be able to draw the cannon in the correct place.

The bullets .x and .y parameters are now being calculated to have the bullet go in a straight line up the screen.