lab 4: Limit cannon movement to screen width.
The set_cannon_boundary checks if the cannon_position is out if bounds and if it tries to be it sets it the boundary itself. Thanks to this the cannon cannot leave the screen.

Additionally the draw_scene function has been renamed to draw_bckg_abd_cannon as "scene" is too general of a description. The background has to be always redrawn before the cannon, else there would be a trace of the cannons past positions.