lab 4: Redesign the explosion animation mechanism.
An array of explosion structs has been introduced, containing the position, animation scale and amount of frames left.

The previous method of animating explosions was flawed as when a new target was hit during the animation of another one, the animation of both started again. The new approach gives every potential target explosion it's own scale and animation time.

As a result of the new approach the code for drawing and keeping track of explosion animations (a particular for loop) has been greatly simplified.

The Target struct has lost the is_exploding, x_boom and y_boom variables and the explosion_frame_counter variable has been removed.