lab 5: Make target bullets not dissapear randomly.
In the previous version the state of should_shoot was independent. This made it possible for the state of bullet visibility to change randomly.
Now it is dependent on bullet visibility.
This means that if a bullet is already visible it's state won't be change randomly.

In addition to this the code responsible for controlling and drawing the bullets has been extracted into a separate for loop.

The chance of a target shooting has also been decreased.

The player_is_hit detection still doesn't register hits.