lab 5: Create hit areas on the player cannon.
The targets' bullets can now hit the player. The hit areas are implemented as 7 hit circles covering the entire cannon.
 If a bullets gets sufficiently close to any of them the player_is_hit function returns 'true'.

In addition the cannon_position has been converted to a coordinate system where the relative '0' is in the middle of the screen.
This has been done as the player bullets and cannon 'y'
already use a relative coordinate system and it makes calculations using combinations of these parameters