lab 5: Create players lives.
MAX_LIVES and STARTING_LIVES have been defined.
The player loses one life for being hit and gains 1 life for hitting 10 targets.
The lives left are drawn in the bottom left corner.

A lives_left variable has been created to keep track of the players lives.

A counter_control vhas been created to control when another life is given to the player.
It is needed as without it if a multiplicity of 10 targets have been hit lives would be added to lives_left until another target has been hit.