lab 5: Add final score to game_over and add possibility to play again.
The game_over function has been modified to now include drawing the final score.

In addition to this now in game_over the player can choose whether to exit or to play again.
To achieve the latter if the player presses space in game_over he breaks out of the loop in said function.
This brings him back to the main game loop where now a goto statement is executed which goes to almost the beggining of main() in order to restart all variables.

Additionally 'stdlib.h' has been included.