lab 5: Iterate the process of shooting bullets.
The code responsible for shooting MAX_BULLETS amount of bullets is now iterated.
This means that the MAX_BULLETS value is now the only thing that has to change to change the amount of bullets.

Earlier if N amount of bullets was wanted the 'if (should_shoot == true)' section of code would have to have N segments one for each bullet.
They would have to be added manually thus changing the amount of available bullets an inconvenience.