lab 6: Temporarily set the max amount of enemies to 3 and make them not come back after death.
When the player shoot down all 3 enemies they don't reappear. This has been acheieved by disabling an enemies visibility when it's shot down.

Because the enemies' bullets are connected to the enemies bullets new bugs had to be fixed.
One of them is bullets of a dead enemy appearing at x=0. Another and major one was the enemies bullet disappearing when he got shot down.
To fix this bug the enemie.visible check has been removed from the 'enemies' bullets loop'.
It used to engulf the whole inside of the loop, now it only appears when deciding whether the enemy should shoot or not.