lab 6: Introduce levels to the game.
A Level struct has been created. It holds the max and current amount of enemies per each level as well as a modifier which is not yet used.
You progress to a new level only when the .current_enemies in a givel level reaches 0;

In most for loops the value of MAX_ENEMIES has been replaced by levels[current_level].max_enemies in order to only perform actions on the enemies from the current level.

The value of reference_time gets a new value each time the player progesses to the next level.
This is done in order to have the enemies appear in relation to the start time of the current level, not the start time of the 'zero' level.

The values for all levels are initialized before the START label as they don't vary between game launches.
The value of .max_enemies in a given level is still limited by the global MAX_ENEMIES.
If a higher values is to be set - it gets set to MAX_ENEMIES.
Because of this the higher levels might have the same amount of enemies.
The .enemy_characteristic can still increase though, at least for now.