lab 6: Go to game_over when the player reaches MAX_LEVEL.
To do this the START label was moved above the levels[] initalization loop to have the amount of .current_enemies reset after finishing the game.

The 'if' responsible for updating the current_level has been moved above the code that checks if the game is over.

Additionally the current_level stat initialization have been moved into the initial Struct Stats stats initialization.

In game_over the frame and title change color to red if you lost and to green if you won.