lab 6: Allow the players bullets to exist on the screen boundary.
Now the players bullets can travell on the screen edges.
Earlier they immediately disappeared when shot on the boundary.

This bug existed because set_ship_boundary allows the center of the ship to exist on the boundary, so when the bullets weren't allowed to be there it looked and felt just wrong.