lab 6: Make enemies spawn on both sides of the screen and bounce from the edges.
The enemies can now spawn on either the right or left edge of the scree at random.
Additionally they bounce from the edges instead of teleporting to the other side.

To do this a new variable 'direction' was added to the Enemy struct.
If direction = 1 the enemy is moving to the right.
If direction = -1 the enemy is moving to the left.

The initial spawn location and correspoding direction are decided in the for loop responsible for the initialization of some enemy and explosion variables. in line 515.

This required some changes int destroy_enemy() as now,  after hitting the enemy, it has to randomly choose where to spawn the enemy and provide the approperiate direction. Earlier it just set the enemies position to x=0;