lab 6: Initialize enemies[].visible.
This one line of code has been missing for a long time. It's lack was the cause of a very irritating bug. It took a lot of time to track it down.

Without this initialization the enemies in a given level seemed to appear correctly however sometimes an extra enemy would appear.
It behaved just like the other ones and if shot down, it did count towards finishing the level.

Lesson learned: ALWAYS INITIALIZE VARIABLES WHICH YOU USE!