lab 6: Introduce transitions between levels.
When a sufficent amount of enemies has been cleared from a level a tranisition to the next level occurs.

A new variable 'LEVEL_TRANSITION_TIME' is defined. It provides the time length of the transition.

The tranisiton itself is created inside the level_transition() function.
It creates a visual copy of the main game universe: It takes the ships position, the stars and the stats.
It also takes frame_time in order to have the fps cap functional in the tranistions.