lab 6: Check if enemy is visible before checking if it's hit.
One more condition was added. An enemy now can be hit only if the bullet is visible, is_hit == true and the enemy itself is visible.

Earlier after hitting an enemy it reset it position to x = 0, stoped being visible, yet could stll be hit causing errors in the enemies/level progression.